:recycle: Suite tests
